Move bridge related vms into single folder
