+++ lib/view_model/bridge/bridge_details_view_model.dart
