+++ lib/view_model/bridge/bridge_history_view_model.dart
